Split logic into logical packages
