--- app/src/main/java/au/com/beba/chainReaction/feature/LocalPreferences.kt
@@ -1,3 +1,0 @@
-package au.com.beba.chainReaction.feature
-
-interface LocalPreferences
