--- app/src/main/java/au/com/beba/chainReaction/feature/LocalPreferencesImpl.kt
@@ -1,5 +1,0 @@
-package au.com.beba.chainReaction.feature
-
-import android.content.Context
-
-class LocalPreferencesImpl(private val context: Context) : LocalPreferences
