+++ app/src/main/java/au/com/beba/chainReaction/feature/chain/ChainCallback.kt
@@ -1,0 +1,12 @@
+package au.com.beba.chainReaction.feature.chain
+
+interface ChainCallback {
+    fun onDone(status: Status)
+
+    enum class Status {
+        NOT_STARTED,
+        IN_PROGRESS,
+        ERROR,
+        SUCCESS
+    }
+}
