+++ app/src/main/java/au/com/beba/chainReaction/feature/chain/ChainDecisionListener.kt
@@ -1,0 +1,6 @@
+package au.com.beba.chainReaction.feature.chain
+
+interface ChainDecisionListener {
+    fun onDecisionDone(finalStatus: ChainCallback.Status)
+    fun onDecisionNotDone()
+}
