+++ app/src/main/java/au/com/beba/chainReaction/feature/chain/Reaction.kt
@@ -1,0 +1,3 @@
+package au.com.beba.chainReaction.feature.chain
+
+class Reaction(val type: String, val task: (Any?) -> (Any?))
