+++ app/src/main/java/au/com/beba/chainReaction/feature/chain/Reactor.kt
@@ -1,7 +1,5 @@
 package au.com.beba.chainReaction.feature.chain
 
-import au.com.beba.chainReaction.feature.reactors.ChainDecision
-
 interface Reactor {
     val chainDecision: ChainDecision
 }
