+++ app/src/main/java/au/com/beba/chainReaction/feature/reactor/withPhases/BaseChainWithPhases.kt
@@ -1,15 +1,16 @@
-package au.com.beba.chainReaction.feature.reactors
+package au.com.beba.chainReaction.feature.reactor.withPhases
 
-import au.com.beba.chainReaction.feature.ConsoleLogger
-import au.com.beba.chainReaction.feature.chain.Reactor
+import au.com.beba.chainReaction.feature.chain.*
+import au.com.beba.chainReaction.feature.logger.ConsoleLogger
+import au.com.beba.chainReaction.feature.reactor.BaseReactor
 
-abstract class BaseChain(private val reactor: Reactor = BaseReactor()) :
+abstract class BaseChainWithPhases(private val reactor: Reactor = BaseReactor()) :
         AbstractChain(reactor),
         ChainWithPhases,
         ChainDecisionListener {
 
     @Suppress("PropertyName")
-    override val TAG: String = BaseChain::class.java.simpleName
+    override val TAG: String = BaseChainWithPhases::class.java.simpleName
 
     private val links: MutableList<Chain> = mutableListOf()
     private val reactions: MutableList<Reaction> = mutableListOf()
